Worked on the FileInput class, expanded my tests and made som imports
I think I managed to import the swing and scalaFX libraries and got a
bit familiar with scalaFX.
Tried to get the source.fromFile to work in fileinput but I haven't
gotten it to work correctly. When testing, it returns strange values. I
have tried it on data.csv and other files as well. Should return 1, but
returns 51.
Also, for some reason, I'm having difficulties creating an object of
type: SimpleRegression.